requirements.txt
pillow
opencv-python
numpy
scikit-learn